Refactor division operations to use native Python 3 semantics

+++ emission/analysis/intake/segmentation/trip_segmentation_methods/dwell_segmentation_dist_filter.py
@@ -1,5 +1,3 @@
-from builtins import str, range, zip
-from past.utils import old_div
 import logging
 import attrdict as ad
 import numpy as np
@@ -96,7 +94,7 @@ class DwellSegmentationDistFilter(eaist.TripSegmentationMethod):
 
         # --- Define thresholds for segmentation ---
         # For a candidate segmentation, if the speed is below this threshold, then the phone may be stationary.
-        speed_threshold = float(self.distance_threshold * 2) / (self.time_threshold / 2)
+        speed_threshold = (self.distance_threshold * 2) / (self.time_threshold / 2)
 
         with ect.Timer() as t_loop:
             # --- Identify candidate segmentation points ---
